@@ -25,12 +25,11 @@ namespace Sample.AspNetCore.SystemTests.Test.PaymentTests.Payment
             Assert.DoesNotThrowAsync(async () =>
             {
 
-                GoToOrdersPage(products, payexInfo, Checkout.Option.LocalPaymentMenu)
-                    .PaymentLink.StoreValueAsUri(out var paymentLink)
-                    .Actions.Rows[y => y.Name.Value.Contains(PaymentResourceOperations.CreateCancellation)].ExecuteAction.ClickAndGo()
-                    .Actions.Rows[y => y.Name.Value.Contains(PaymentResourceOperations.PaidPayment)].Should.BeVisible()
-                    .Actions.Rows[y => y.Name.Value.Contains(PaymentResourceOperations.ViewPayment)].Should.BeVisible()
-                    .Actions.Rows.Count.Should.Equal(2);
+	            GoToOrdersPage(products, payexInfo, Checkout.Option.LocalPaymentMenu)
+		            .PaymentLink.StoreValueAsUri(out var paymentLink)
+		            .Actions.Rows[y => y.Name.Value.Contains(PaymentResourceOperations.CreateCancellation)].ExecuteAction.ClickAndGo()
+		            .Actions.Rows[y => y.Name.Value.Contains(PaymentResourceOperations.PaidPayment)].Should.BeVisible()
+		            .Actions.Rows[y => y.Name.Value.Contains(PaymentResourceOperations.ViewPayment)].Should.BeVisible();
 
                 var cardPayment = await SwedbankPayClient.Payments.CardPayments.Get(paymentLink, PaymentExpand.All);
 
